% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{als_data}
\alias{als_data}
\title{ALS Example dataset}
\format{
A list
\describe{
\item{subjid}{Subject ID}
\item{p1}{ALSFRS-R 1}
\item{p2}{ALSFRS-R 2}
\item{p3}{ALSFRS-R 3}
\item{p4}{ALSFRS-R 4}
\item{p5}{ALSFRS-R 5}
\item{p6}{ALSFRS-R 6}
\item{p7}{ALSFRS-R 7}
\item{p8}{ALSFRS-R 8}
\item{p9}{ALSFRS-R 9}
\item{x1r}{ALSFRS-R R1}
\item{x2r}{ALSFRS-R R2}
\item{x3r}{ALSFRS-R R3}
\item{age_at_baseline}{Age at baseline}
\item{age_at_onset}{Age at onsite}
\item{onset}{Region of onset}
\item{baseline_date}{Baseline date3}
\item{death_date}{Death date}
}
}
\usage{
als_data
}
\description{
An ALS-related example dataset.
}
\keyword{datasets}
