% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{als_data_qc_mapping}
\alias{als_data_qc_mapping}
\title{An example dataset containing a Quality Control mapping}
\format{
A list of 3 \code{tibbles}.

\itemize{
\item{missing}{Table with all the 'missing' tests.  }
\item{inconsistencies}{Table with all the 'inconsistencies' tests.}
\item{range}{Table with all the 'out of range' tests.}
}
}
\usage{
als_data_qc_mapping
}
\description{
An example dataset containing a Quality Control mapping
}
\keyword{datasets}
