% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{AQFactor}
\alias{AQFactor}
\title{Function to replace Absent and Questionable levels with A/Q.}
\usage{
AQFactor(x)
}
\arguments{
\item{x}{character vector or factor}
}
\description{
Takes a character vector or factor and returns a vector where Absent and Questionable entries are replaced by A/Q.
}

