% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_modals}
\alias{convert_modals}
\title{Convert Modalities}
\usage{
convert_modals(modals, output = c("New Labels", "Modality"),
  conversion_document = "K:/Stat/RMT/Efficiency Reports/Data/modalities_conversion.xlsx",
  sheet = "modals_only")
}
\arguments{
\item{modals}{Modalities to convert}

\item{output}{Any combination of 'New Labels' and 'Modality'}

\item{conversion_document}{Document from which the conversion matrix should be loaded.}

\item{sheet}{If not first sheet in conversion_document, then specify here}
}
\description{
Takes any modality from any of the modalities listed in the spreadsheet provided and
spits out any combination of the following: New Label, Eyes Graded, Minutes.
}
