% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_names}
\alias{convert_names}
\title{Convert Names}
\usage{
convert_names(names, out_format = c("Full Name", "First Name", "Last Name",
  "Number"), conversion_document = "K:/Stat/RMT/grader_names.xlsx",
  load = TRUE, ig.case = TRUE)
}
\arguments{
\item{names}{Names to convert}

\item{out_format}{Should names be returned as Full Name, First Name, Last Name, or Number}

\item{conversion_document}{either matrix with two columns or path to document containing conversion
scheme. (If document, remember 'load = TRUE'.) Should consist of two columns, first with first and
last names separated by space, second with everything to match against.}

\item{load}{logical; if TRUE, conversion_document is assumed to be an excel spreadsheet and will be read using read_excel from the readxl package.}

\item{ig.case}{ignore case when matching?}
}
\description{
Take any name from any of the given naming conventions and
spit out the wanted name.
}
\examples{
\dontrun{
convert_names(names = c('pak', 'sheila'),
              out_format = c('Full Name'),
              conversion_document = 'K:/Stat/RMT/grader_names.xlsx',
              load = TRUE,
              ig.case = TRUE)}
}
