% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossa.R
\name{crossa}
\alias{crossa}
\title{Cross Tabs with Shading Algorithms}
\usage{
crossa(data1a, lab1, data2a, lab2, cantgrade, title1 = "", footer = "",
  title2 = "", ext.labels = NULL)
}
\arguments{
\item{lab1}{label for data1}

\item{lab2}{label for data2}

\item{cantgrade}{vector containing values that are used for cannot grade and hence
should not be included in within one step calculations.}

\item{title1}{main title of graph, line one}

\item{footer}{footer for graph}

\item{title2}{main title of graph, line two}

\item{ext.labels}{if NULL (default), labels printed in squares. Otherwise, character
vector with same length as there are levels of the variable considered, with strings
to be stand-in for labels in squares while actual labels are printed in the lower left
corner. If true, numbers are used as labels and automatically}

\item{data1}{a vector of grades from grader 1; not a factor}

\item{data2}{a vector of grades from grader 2; not a factor}
}
\description{
This functions allows you to create cross tabs with
shading algorithms and fills it with exact agreement,
agreement within two steps, kappa, the standard error of kappa,
weighted kappa (within one step), and the standard error
of weighted kappa.
}
\examples{
\dontrun{
crossa(data1a = FIDI$CNVPRS.F,
       lab1   = "FILM SYNTHETIC GRADES",
       data2a = FIDI$CNVPRS.D,
       lab2   = "DIGITAL SYNTHETIC GRADES",
       title1 = "Presence of Classic CNV",
       footer = "Comparison of Digital Synthetic Grades and
                Film Synthetic Grades")
                }
}

