% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.agreement.R
\name{exact.agree.table}
\alias{exact.agree.table}
\title{Table with Exact Agreement}
\usage{
exact.agree.table(data, crossA, plotX, grads, nes = c("", NA),
  cantgrade.crossA = NULL, cantgrade.plotX = NULL)
}
\arguments{
\item{data}{Data to use}

\item{crossA}{vector with variables used for crossA tables. If named, names are used as variable labels}

\item{plotX}{vector with variables used for twogradersplotx. If named, names are used as variable labels}

\item{grads}{vector with names or graders}

\item{nes}{vector with character strings to exclude}
}
\description{
Calculates agreement table for overall team and for graders
individually.
}
