% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frontpage.tables.R
\name{grader.frontpage.table}
\alias{grader.frontpage.table}
\title{Table for Front Page of Grader Specific Report}
\usage{
grader.frontpage.table(data, QCST = c("ST", "QC"))
}
\arguments{
\item{data}{data set used for analysis}

\item{QCST}{use QC or ST dates?}
}
\description{
Creates table with four columns: QC Grader, N, QC Grading Start Date, QC Grading End Date
}
