% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ICC}
\alias{ICC}
\title{Interclass Correlation}
\usage{
ICC(grade1, grade2, digits = 3)
}
\arguments{
\item{grade1}{vector containing gradings done by grader 1}

\item{grade2}{vector containing gradings done by grader 2}

\item{digits}{number of digits to round to; default 3}
}
\description{
Calculates interclass correlation between two graders.
}

