% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{load.file}
\alias{load.file}
\title{Load file from data.dir}
\usage{
load.file(data.dir, file)
}
\arguments{
\item{data.dir}{data directory in which the file is located}

\item{file}{full name of file to be loaded (i.e. of the form 'filename.xlsx')}
}
\description{
Loads the file called `file` from folder `data.dir` and creates `Project` variable with value filename
}
