% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-CrossA.R
\name{Plot.crossA}
\alias{Plot.crossA}
\title{Wrapper for crossA}
\usage{
Plot.crossA(data, crossAs = NULL, vars = "all", lab1 = "ST",
  lab2 = "QC", cannotgrades = NULL, foottext = NULL, show.plot = FALSE,
  sep = ".", nes = c("", NA), title2 = "", ext.labels = NULL)
}
\arguments{
\item{data}{dataset that includes columns named <var>.QC and <var>.ST for all <var> in crossAs}

\item{crossAs}{named vector containing variables from
data. Names are used as titles for
the plots. If not named, the variable itself
will be used.}

\item{vars}{indicates which variables should be plotted. Default is 'all'}

\item{lab1}{lab1 in crossa (default 'ST')}

\item{lab2}{lab2 in crossa (default 'QC')}

\item{cannotgrades}{Named vector indicating how 'cannot grade' is
coded. Names should be identical to names(crossAs)}

\item{foottext}{foot text for plots (default NULL)}

\item{show.plot}{logical; should the plots be plotted one by one? (click for next, default FALSE)}

\item{sep}{separation symbol between variable and QC/ST in data. (default '.')}

\item{nes}{character vector with strings that should be ignored in right.subset. (default c('', NA))}

\item{title2}{character string to use as subtitle for plots}

\item{ext.labels}{if NULL (default), labels printed in squares. Otherwise, character
vector with same length as there are levels of the variable considered, with strings
to be stand-in for labels in squares while actual labels are printed in the lower left
corner. If true, numbers are used as labels and automatically.}

\item{angle}{numeric; should variable categories be angled when added?}
}
\description{
For a given dataset and a named vector with
variables from that dataset, this function creates
crossa plots for all variables.
}

