% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-PlotxsdLimit.R
\name{Plot.plotx.sd.Limit}
\alias{Plot.plotx.sd.Limit}
\title{Wrapper for twograders.plotx and sd.plotaLimit}
\usage{
Plot.plotx.sd.Limit(data = NULL, plotX = NULL, vars = "all",
  lab1 = "ST", lab2 = "QC", sep = ".", nes = c("", NA), ranges = c(1,
  2, 5), cantgrade = 8888, foottext = "", title2 = "",
  show.plot = FALSE)
}
\arguments{
\item{data}{dataset}

\item{plotX}{named vector containing variables from
data. Names are used as titles for
the plots. If not named, the variable itself
will be used.}

\item{vars}{indicates which variables should be plotted. Default is 'all'}

\item{lab1}{lab1 in crossa (default 'ST')}

\item{lab2}{lab2 in crossa (default 'QC')}

\item{sep}{separation symbol between variable and QC/ST in data. (default '.')}

\item{nes}{character vector with strings that should be ignored in right.subset. (default c('', NA))}

\item{foottext}{foot text for plots (default NULL)}

\item{title2}{character string to use as subtitle for plots}

\item{show.plot}{logical; should the plots be plotted one by one? (click for next, default FALSE)}
}
\description{
For a given dataset and a named vector with
variables from that dataset, this function creates
twograders.plotx and sd.plotaLimit (when appropriate) plots
for all variables.
}

