% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{recode_DRsev}
\alias{recode_DRsev}
\title{Recode DR Severity Levels}
\usage{
recode_DRsev(DRSeverity)
}
\arguments{
\item{DRSeverity}{vector of severity levels to be recoded}
}
\description{
Takes a vector of DR Severity levels and returns the recoded values. 'Not applicable' is turned into NA.
}
