% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{right.subset.R}
\alias{right.subset.R}
\title{Right Subset}
\usage{
right.subset.R(x, var, out = c("ST", "QC"), sep = ".", nes = "")
}
\arguments{
\item{x}{data}

\item{var}{variable to select}

\item{out}{do you want ST, QC or both ('ST', 'QC')?}

\item{sep}{how are variables separated from ST/QC specification?}

\item{nes}{character vector specifying if any strings should be excluded.}
}
\description{
Function to select the right subset of the data
}

