% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exact.agreement.R
\name{save.exact}
\alias{save.exact}
\title{Save Outcome of exact.agreement}
\usage{
save.exact(exact.agreement, path, file)
}
\arguments{
\item{exact.agreement}{object returned by exact.agree.table}

\item{path}{character string specifying path for files to be saved to}

\item{file}{name of file containing all data.}
}
\description{
save.exact writes the results of exact.agree.table to excel spreadsheets. One for each grader together with one containing everything.
}
