% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sd.plotaLimit.R
\name{sd.plotaLimit}
\alias{sd.plotaLimit}
\title{Bland-Altman Plot with Confidence Interval}
\usage{
sd.plotaLimit(grade1a, grade2a, label1, label2, range1, range2, range3,
  ptsz = 0.7, cantgrade = "", maintitle = "", subtitle = "",
  footer = "", x1 = -999, x2 = -999, y1 = -999, y2 = -999,
  factor = 1e-08, spline = FALSE)
}
\arguments{
\item{grade1a}{a vector of grades}

\item{grade2a}{a corresponding vector of grades}

\item{label1}{label of the first grader}

\item{label2}{label of the second grader}

\item{range1}{numeric}

\item{range2}{numeric}

\item{range3}{numeric}

\item{cantgrade}{character/numeric used to indicate cannot grade}

\item{maintitle}{main title of graph}

\item{subtitle}{sub title of graph}

\item{footer}{footer}

\item{x1, x2}{range of x axis -- if -999, the range is calculated based on the data}

\item{y1, y2}{range of y axis -- if -999, the range is calculated based on the data}

\item{factor}{numeric; passed to jitter for plotting of data points. Adjusts how spread out the points are.}
}
\description{
Creates a plot of the difference between two grades against
the mean of the two grades with smooth mean and smoothed 95%
bound for continuous type data.
Calculates the mean difference, intra-class correlation,
within [range1](\%) and within [range2](\%)
}

