% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_function.R
\name{test_inconsistencies}
\alias{test_inconsistencies}
\title{Test the inconsistencies between variables on a dataset}
\usage{
test_inconsistencies(data, variable1, variable2, relation)
}
\arguments{
\item{data}{data to be tested.}

\item{variable1}{The variable to be tested.}

\item{variable2}{The variable to be tested.}

\item{relation}{String such as 'greater_than', 'greater_than_or_equal'
'lower_than_or_equal' and 'lower_than'.}
}
\value{
A data frame containing all the findings regarding the applied test.
}
\description{
Test the inconsistencies between variables on a dataset
}
\examples{
test_inconsistencies(als_data, 'baseline_date', 'death_date', relation = 'lower_than')
test_inconsistencies(als_data, 'age_at_baseline', 'age_at_onset', relation = 'greater_than')
}
