% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_function.R
\name{test_missing}
\alias{test_missing}
\title{Test the variable missingness on a dataset}
\usage{
test_missing(data, variable)
}
\arguments{
\item{data}{data to be tested.}

\item{variable}{The variable to be tested.}
}
\value{
A data frame containing all the findings regarding the applied test.
}
\description{
Test the variable missingness on a dataset
}
\examples{
test_missing(als_data, 'p8')
test_missing(als_data, 'p1')
}
