% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twograders.plotx.R
\name{twograders.plotx}
\alias{twograders.plotx}
\title{Correlation Plot}
\usage{
twograders.plotx(grade1a, grade2a, label1, label2, range1, range2,
  range3 = 9999, ptsz = 0.7, lab = c(5, 5, 7), cantgrade = "",
  maintitle = "", subtitle = "", footer = "", shiny = FALSE,
  x1 = -999, x2 = -999, y1 = -999, y2 = -999, mar = c(0, 0, 0, 0))
}
\arguments{
\item{grade1a}{vector of grades}

\item{grade2a}{vector of second grades}

\item{label1}{label for first grades}

\item{label2}{label for second grades}

\item{range1, range2, range3}{numerics; specify the ranges to use for within calculations}

\item{ptsz}{numeric; passed to cex argument of plot}

\item{cantgrade}{numeric value specifying what is used for 'Cannot Grades'}

\item{maintitle}{main title of graph}

\item{subtitle}{sub title of graph}

\item{footer}{foot text}

\item{x1, x2}{range for x axis -- if -999 the range is calculated from data}

\item{y1, y2}{range for y axis -- if -999 the range is calculated from data}
}
\description{
Creates the plot with x-axis = "Grader 1" vs. Y-axis = "Grader 2" for
continuous type data.
Calculates the mean difference, intra-class correlaion, within [range1] (\%),
and within [range2](\%)
}
