% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unfolded.kappa.R
\name{unfolded.kappa}
\alias{unfolded.kappa}
\title{Unfolded Kappa}
\usage{
unfolded.kappa(x, y, lev = unique(c(levels(x), levels(y))), nlev = NULL,
  wgt = NULL, cantgrade = cantgrade)
}
\arguments{
\item{x}{vector of first grades}

\item{y}{vector of second grades}

\item{lev}{vector with levels of x and y grades}

\item{nlev}{number of levels. If NULL, length of lev is used}

\item{wgt}{weights to be used. If NULL, nlev is used}

\item{cantgrade}{string which specifies 'Cannot Grade'}
}
\description{
Caculates the kappa statistic with an option of weights
}

