% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_sentence1}
\alias{wrap_sentence1}
\title{Wrap Sentence}
\usage{
wrap_sentence1(string, width)
}
\arguments{
\item{string}{sentence to wrap}

\item{width}{wanted width of outcome}
}
\description{
This is used to wrap long level names of factors for plotting of crossA plots.
It will first split according to commas, then according to words.
}

