% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-CrossA.R
\name{Plot.crossA}
\alias{Plot.crossA}
\title{Wrapper for crossA}
\usage{
Plot.crossA(data, crossAs = NULL, vars = "all", lab1 = "ST",
  lab2 = "QC", cannotgrades = NULL, foottext = NULL, show.plot = FALSE,
  show.progress = TRUE, sep = ".", nes = c("", NA), title2 = "",
  ext.labels = NULL, cex = 1, PLOT = TRUE, mar = c(0, 0, 0, 0))
}
\arguments{
\item{data}{dataset that includes columns named <var>.QC and <var>.ST for all <var> in crossAs}

\item{crossAs}{named vector containing variables from
data. Names are used as titles for
the plots. If not named, the variable itself
will be used.}

\item{vars}{indicates which variables should be plotted. Default is 'all'}

\item{lab1}{lab1 in crossa (default 'ST')}

\item{lab2}{lab2 in crossa (default 'QC')}

\item{cannotgrades}{Named vector indicating how 'cannot grade' is
coded. Names should be identical to names(crossAs)}

\item{foottext}{foot text for plots (default NULL)}

\item{show.plot}{logical; should the plots be plotted one by one? (click for next, default FALSE)}

\item{sep}{separation symbol between variable and QC/ST in data. (default '.')}

\item{nes}{character vector with strings that should be ignored in right.subset. (default c('', NA))}

\item{title2}{character string to use as subtitle for plots}

\item{ext.labels}{if NULL (default), labels printed in squares. Otherwise, character
vector with same length as vars, with strings to be stand-in for labels in squares while actual labels
are printed in the lower left corner. If elements are not strings, but TRUE/FALSE, numbers are used as labels
for TRUE entries. If simply TRUE, external labels will be used for all variables.}

\item{cex}{numeric; passed to crossa. Default 1.}

\item{PLOT}{logical; if TRUE (default), figure is plotted. If FALSE, summary returned without any figures.}
}
\description{
For a given dataset and a named vector with
variables from that dataset, this function creates
crossa plots for all variables.
}
